refactor: 비속어 필터 이동
