+++ src/main/java/kitchenpos/profanity/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.profanity.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
